Prefix models with 'API Models/' and refactor APIs
Normalize model naming to include an "API Models/{provider}/" prefix (ModelDefinition.GetFullName and provider lookup/normalization in APIProviderInit/DynamicAPIBackend). Update backend URL selection to use the new prefix and simplify some ideogram/google URL logic. Change Fal seed type to long across builders and params, adjust OutputFormat/other param aliases to reuse existing definitions, and switch model Tags to a string[] type. Add project file, .gitignore entry for /obj, and perform general using/order/formatting cleanup and small refactors in SwarmUIAPIBackends (new Ideogram params, consolidated group initialization, and condensed flag registration).

+++ Providers/RequestBuilders.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
+using SwarmUI.Backends;
 using SwarmUI.Text2Image;
 using SwarmUI.Utils;
 using Hartsy.Extensions.APIBackends.Models;
@@ -458,7 +459,7 @@ namespace Hartsy.Extensions.APIBackends.Providers
 
             request["num_images"] = GetNumImages(input);
 
-            if (input.TryGet(SwarmUIAPIBackends.SeedParam_Fal, out int seed) && seed >= 0)
+            if (input.TryGet(SwarmUIAPIBackends.SeedParam_Fal, out long seed) && seed >= 0)
                 request["seed"] = seed;
 
             if (input.TryGet(SwarmUIAPIBackends.GuidanceScaleParam_Fal, out double guidance))
